+++ Sources/PreludeAuth/HTTP/Headers.swift
@@ -15,4 +15,5 @@ enum HTTPHeader {
     static let deviceID = "X-Device-Id"
     static let refreshToken = "X-Refresh-Token"
     static let refreshTokenExpiresAt = "X-Refresh-Token-Expires-At"
+    static let verificationToken = "X-Verification-Token"
 }